+++ crates/ruma-common/tests/events/ui/07-enum-sanity-check.rs
@@ -1,15 +1,16 @@
+use ruma_common::events;
 use ruma_macros::event_enum;
 
 event_enum! {
     /// Any global account data event.
     enum GlobalAccountData {
-        "m.direct",
+        "m.direct" => events::direct,
         #[cfg(test)]
-        "m.ignored_user_list",
+        "m.ignored_user_list" => events::ignored_user_list,
         // Doesn't actually have a wildcard, but this should work as a wildcard test
-        "m.push_rules.*",
+        "m.push_rules.*" => events::push_rules,
         #[cfg(any())]
-        "m.ruma_test",
+        "m.ruma_test" => events::ruma_test,
     }
 }