+++ crates/ruma-client-api/src/membership.rs
@@ -54,9 +54,8 @@ impl<'a> ThirdPartySigned<'a> {
 ///
 /// To create an instance of this type, first create a `Invite3pidInit` and convert it via
 /// `Invite3pid::from` / `.into()`.
-#[derive(Clone, Debug, PartialEq, Incoming, Serialize)]
+#[derive(Clone, Debug, Incoming, Serialize)]
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
-#[incoming_derive(PartialEq)]
 pub struct Invite3pid<'a> {
     /// Hostname and port of identity server to be used for account lookups.
     pub id_server: &'a str,