+++ crates/ruma-client-api/CHANGELOG.md
@@ -1,5 +1,10 @@
 # [unreleased]
 
+Improvements:
+
+* Add `From<&UserId>` and `From<&OwnedUserId>` implementations for `UserIdentifier`
+* Add `UserIdentifier::email` constructor
+
 # 0.14.0
 
 Bug fixes: