+++ crates/ruma-client-api/CHANGELOG.md
@@ -1,5 +1,10 @@
 # [unreleased]
 
+Breaking changes:
+
+* Remove `PartialEq` implementations for a number of types
+  * If the lack of such an `impl` causes problems, please open a GitHub issue
+
 Improvements:
 
 * Add `From<&UserId>` and `From<&OwnedUserId>` implementations for `UserIdentifier`